Extract each handler to its own file
