+++ test/demo-extension/background/sum.ts
@@ -1,0 +1,8 @@
+import {Contract} from '../../../index';
+
+export const sumContract: Contract<typeof sum> = {
+  type: 'sum',
+};
+export async function sum(...addends: number[]): Promise<number> {
+  return addends.reduce((a, b) => a + b);
+}
